[DOP-20067] move RunStatus enum from types to constants
