+++ src/entities/run/api/constants.ts
@@ -1,0 +1,8 @@
+export enum RunStatus {
+  CREATED = 'CREATED',
+  STARTED = 'STARTED',
+  FAILED = 'FAILED',
+  SEND_STOP_SIGNAL = 'SEND STOP SIGNAL',
+  STOPPED = 'STOPPED',
+  FINISHED = 'FINISHED',
+}
