+++ src/entities/run/api/types.ts
@@ -1,5 +1,7 @@
 import { PaginationRequest } from '@shared/types';
 
+import { RunStatus } from './constants';
+
 export interface Run {
   id: number;
   transfer_id: number;
@@ -7,15 +9,6 @@ export interface Run {
   ended_at: string | null;
   status: keyof typeof RunStatus;
   log_url: string | null;
-}
-
-export enum RunStatus {
-  CREATED = 'CREATED',
-  STARTED = 'STARTED',
-  FAILED = 'FAILED',
-  SEND_STOP_SIGNAL = 'SEND STOP SIGNAL',
-  STOPPED = 'STOPPED',
-  FINISHED = 'FINISHED',
 }
 
 export interface GetRunsRequest extends PaginationRequest {
